Add download status to statusbar
It overlaps the first status field but is only visible while downloading
and shortly after.

Active downoads are indicated by yellow background, a text and a
progress bar.

If all downloads finish, the bar turns greend and disappers after 5 sec.

The data is collected from DownloadObserver and uses a wxTimer to get
updates.

see  Issue #49