fix #119 Scroll bars everywhere (windows)
at least for battlelist and replays

this had something todo with the initial width of the listctrl
I could not track down to which change exactly caused the issue
but setting a fixed width before calling Layout on any parent
causes the parent ho have the scrollbar.

wxLIST_AUTOSIZE only works if the columns have some content,
so we have to call it when adding an item

SetColumnWidth does more than it says: it will set the specified column width
or load the column width if set anytime before.