refactor path handeling
Sorry for the big commit. I could not find a good point to stop with
a working springlobby.

This commit changes the way paths are handled. It does not fix
any issues yet, but is a step.

- move functions which determine used paths to SlPaths
  or remove if unused
- configure LSL paths trough LSL::Util::config().ConfigurePaths(),
  used LSL version does not relay on Settings() any more

The commit touches many files, but most of them because of
the change from "sett()." to "SlPaths::"