Added wx-config version selector on fail.
Append version selector (major+minor) to wx-config calls to find the right one in a multi-slot install if the initial search was unsuccessful. The configuration might otherwise fail if the required version is not the system default. This still fails if a higher (compatible) version than WX_VERSION_MIN is available, but WX_VERSION_MIN is not.